jail: Keep the parent domain root directory in a hidden place
For now, this directory can be accessed through the "." source with the
cmd/mount command from inside the jail.

/!\ WARNING: ZSH seems to have a bug because it behave differently when
listing an absolute or a relative path. Changing directory into the
parent domain seems to be the same as changing directory to the jail
root, but it is not! Maybe the builtin "ls" use an inode cache?