stemjail: Replace the LD_PRELOAD environment with /etc/ld.so.preload
Require a standalone stemshim library.

This library's path must be referenced by the STEMJAIL_LIB_SHIM_PATH
environment variable, but only on the Portal side.

For now, the host must have an existing /etc/ld.so.preload file (needed
to mount onto, in a jail), but it can be empty.